requirements.txt
python-dotenv
python-telegram-bot
nltk
lingua 
lingua-language-detector